+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 0.6.6 (Unreleased)
+## 0.6.7 (Unreleased)
 
 ### Features Added
 
@@ -10,9 +10,21 @@
 
 ### Bugs Fixed
 
+### Other Changes
+
+## 0.6.6 (2026-04-01)
+
+### Features Added
+
+- Added support for CODEOWNERS "Section" in Label Owners (defaults to "Client Libraries")
+
+### Bugs Fixed
+
 - Fixed sample translation to preserve source directory structure when writing translated files
 
 ### Other Changes
+
+- CODEOWNERS generator supports file paths and doesn't assume all paths are directories
 
 ## 0.6.5 (2026-03-27)